add router guard for api and change response from list to json

+++ routes/posts.mjs
@@ -13,36 +13,57 @@ routerPosts.get("/api/posts", (req, res) => {
     else if (req.query.type) {
         query = {type: req.query.type};
     }
+    else if (req.query.user) {
+        query = {user: req.session.userId}
+    }
     postService.getPosts(query).then(result => {
-        res.send(result);
+        res.send({posts: result});
     });
 });
 
+routerPosts.all("/api/post/*", (req, res ,next) => {
+    if(!req.session.userId) { 
+        res.send({message: "unauthorised call"});
+    }
+    else{
+        next();
+    }
+});
 routerPosts.get("/api/posts/:id", (req, res) => {
-    postService.getPosts(req.params.id).then(result => res.send(result)); 
+    postService.getPosts(req.params.id)
+        .then(result => {    
+            res.send({post: result.length === 1? result[0] : null})}); 
 });
 
 routerPosts.post("/api/post", (req, res) => {
+    // should make a check that we are recieving only post things
     const post = {
      ...req.body,
         user: req.session.userId,
     };
     postService.createPost(post).then(() => {
-        res.redirect("/");
+        res.redirect("/"); //maybe should return a json object and the redirect will heppend from the public folder
     });
 });
 
 routerPosts.patch("/api/post", (req, res) => {
     postService.updatePost(req.query.id, req.body).then(() => {
-        res.redirect("/");
+        res.redirect("/"); //maybe should return a json object and the redirect will heppend from the public folder
     });
 });
 
 routerPosts.delete("/api/post", (req, res) => {
+    if(!req.session.userId) { 
+        res.send({message: "unauthorised call"})
+    }
+    else{ 
+        postService.deletePost(req.query.id).then(() => {
+            res.redirect("/"); //maybe should return a json object and the redirect will heppend from the public folder
+        });
+    };
+
+
     //Either req.query or req.params
-    postService.deletePost(req.query.id).then(() => {
-        res.redirect("/");
-    });
 });
 
 export default routerPosts;
